+++ docassemble_base/docassemble/base/functions.py
@@ -494,18 +494,17 @@ def interview_url(**kwargs):
     url += '?' + '&'.join(map((lambda (k, v): str(k) + '=' + urllib.quote(str(v))), args.iteritems()))
     return url
 
-def set_title(title=None, logo=None, short=None, subtitle=None, tab=None, exit_link=None, exit_label=None):
+def set_title(**kwargs):
     """Sets the title of the interview that is displayed in the navigation
     bar and other places, as well as other features of the navigation bar.
 
     """
-    this_thread.internal['title'] = title
-    this_thread.internal['logo'] = logo
-    this_thread.internal['short title'] = short
-    this_thread.internal['subtitle'] = subtitle
-    this_thread.internal['tab title'] = tab
-    this_thread.internal['exit link'] = exit_link
-    this_thread.internal['exit label'] = exit_label
+    this_thread.internal['short title'] = kwargs.get('short', None)
+    this_thread.internal['tab title'] = kwargs.get('tab', None)
+    for key, val in kwargs.iteritems():
+        key = re.sub(r'_', r' ', key)
+        if key in ('title', 'logo', 'subtitle', 'exit link', 'exit label'):
+            this_thread.internal[key] = val
 
 class DATagsSet():
     def add(self, item):